moved all usage examples to py9b/examples
