+++ py9b/examples/m365_hax.py
